LLM-VM/archive/prelimiary/raw_llm_structrued/requirements.txt
openai>=1.0.0
pyyaml>=6.0



